refactor: ComplexForm 분리하기
