+++ src/components/index.ts
@@ -1,4 +1,5 @@
 import { Header } from "./Header";
 import { ItemList } from "./ItemList";
+import { ComplexForm } from "./ComplexForm";
 
-export { Header, ItemList };
+export { Header, ItemList, ComplexForm };
